pyproject.toml
[tool.poetry]
name = "tutoroffice"
version = "0.1.0"
description = ""
authors = ["Takeshi <orwellj@mail.ru>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.11"
django = "^4.1.6"
djangorestframework = "^3.14.0"
psycopg2 = "^2.9.5"
python-dotenv = "^0.21.1"
pillow = "^9.4.0"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
